began python allocator, haven't tested
I'm being cautious as cygwin's vim is real vim, not the git vim that mplays much nicer.

For them moment, this will check by line and assign registers until they are exhausted. I n retrospect I would have held off on this commit but , as noted, I started the commit and can't be sure that I'll be able to erase everything in the commit message to abort ot. nnngh